remove HermesCounterWithRate
