+++ hermes-common/src/main/java/pl/allegro/tech/hermes/common/metric/SubscriptionHermesCounter.java
@@ -3,9 +3,9 @@ package pl.allegro.tech.hermes.common.metric;
 import com.codahale.metrics.Meter;
 import io.micrometer.core.instrument.Counter;
 import pl.allegro.tech.hermes.api.SubscriptionName;
-import pl.allegro.tech.hermes.metrics.counters.DefaultHermesCounterWithRate;
+import pl.allegro.tech.hermes.metrics.counters.MeterBackedHermesCounter;
 
-public class SubscriptionHermesCounter extends DefaultHermesCounterWithRate {
+public class SubscriptionHermesCounter extends MeterBackedHermesCounter {
 
     private final String graphiteName;
     private final SubscriptionName subscription;
