+++ hermes-frontend/src/main/java/pl/allegro/tech/hermes/frontend/metric/CachedTopic.java
@@ -6,11 +6,11 @@ import pl.allegro.tech.hermes.common.kafka.KafkaTopics;
 import pl.allegro.tech.hermes.common.metric.MetricsFacade;
 import pl.allegro.tech.hermes.common.metric.timer.StartedTimersPair;
 import pl.allegro.tech.hermes.metrics.HermesCounter;
-import pl.allegro.tech.hermes.metrics.HermesCounterWithRate;
 import pl.allegro.tech.hermes.metrics.HermesHistogram;
 import pl.allegro.tech.hermes.metrics.HermesRateMeter;
 import pl.allegro.tech.hermes.metrics.HermesTimer;
 import pl.allegro.tech.hermes.metrics.HermesTimerContext;
+import pl.allegro.tech.hermes.metrics.counters.MeterBackedHermesCounter;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -36,8 +36,8 @@ public class CachedTopic {
     private final HermesHistogram topicMessageContentSize;
     private final HermesHistogram globalMessageContentSize;
 
-    private final HermesCounterWithRate topicThroughputMeter;
-    private final HermesCounterWithRate globalThroughputMeter;
+    private final MeterBackedHermesCounter topicThroughputMeter;
+    private final MeterBackedHermesCounter globalThroughputMeter;
 
     private final HermesCounter published;
