--- hermes-metrics/src/main/java/pl/allegro/tech/hermes/metrics/HermesCounterWithRate.java
@@ -1,4 +1,0 @@
-package pl.allegro.tech.hermes.metrics;
-
-public interface HermesCounterWithRate extends HermesCounter, HermesRateMeter {
-}
